% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.History.R
\docType{data}
\name{Chicago.forecast.10day}
\alias{Chicago.forecast.10day}
\title{Chicago 10-day Forecast}
\format{\preformatted{List of 12
 $ KILCHICA166:List of 10
  ..$ model    :List of 18
  .. ..$ loglik    : num -700
  .. ..$ aic       : num 1409
  .. ..$ bic       : num 1426
  .. ..$ aicc      : num 1409
  .. ..$ mse       : num 0.0192
  .. ..$ amse      : num 0.0478
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 1401
  .. .. ..$ par    : num [1:4] 0.873 0.2 16.541 -0.116
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 119
  .. ..$ residuals : Time-Series [1:617] from 1 to 48.4: 0.00121 -0.00779 0.00446 -0.00398 0.00993 ...
  .. ..$ fitted    : Time-Series [1:617] from 1 to 48.4: 16.4 16.3 16.1 16 15.8 ...
  .. ..$ states    : mts [1:618, 1:2] 16.5 16.4 16.2 16.1 16 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.923 48.385 13
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:4] 0.873 0.2 16.541 -0.116
  .. .. ..- attr(*, "names")= chr [1:4] "alpha" "beta" "l" "b"
  .. ..$ m         : num 13
  .. ..$ method    : chr "ETS(M,A,N)"
  .. ..$ components: chr [1:4] "M" "A" "N" "FALSE"
  .. ..$ call      : language ets(y = x, model = etsmodel, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 16.541 -0.116
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 1.83e-05
  .. ..$ x         : Time-Series [1:617] from 1 to 48.4: 16.4 16.2 16.2 16 16 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:26] from 48.5 to 50.4: 45.6 46.2 46.9 47.5 48.2 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:617] from 1 to 48.4: 16.5 16.2 16.2 16 16 15.8 15.8 15.6 15.6 15.4 ...
  ..$ upper    : mts [1:26, 1:2] 45.8 46.6 47.4 48.2 49 ...
  .. ..- attr(*, "tsp")= num [1:3] 48.5 50.4 13
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  ..$ lower    : mts [1:26, 1:2] 45.3 45.8 46.3 46.8 47.3 ...
  .. ..- attr(*, "tsp")= num [1:3] 48.5 50.4 13
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  ..$ fitted   : Time-Series [1:617] from 1 to 48.4: 16.5 16.3 16.1 16.1 15.8 ...
  ..$ method   : chr "STL +  ETS(M,A,N)"
  ..$ residuals: Time-Series [1:617] from 1 to 48.4: 0.0198 -0.1271 0.0718 -0.0638 0.1572 ...
  ..$ seasonal : Time-Series [1:13] from 47.5 to 48.4: -0.0813 -0.1777 -0.1351 -0.1731 -0.2091 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA130:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -2429
  .. ..$ aic       : num 4868
  .. ..$ bic       : num 4894
  .. ..$ aicc      : num 4868
  .. ..$ mse       : num 0.0368
  .. ..$ amse      : num 0.123
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 4858
  .. .. ..$ par    : num [1:5] 1 0.44 0.849 12.442 -0.137
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 267
  .. ..$ residuals : Time-Series [1:1265] from 1 to 422: 0.0748 0.171 -0.2036 0.073 -0.0653 ...
  .. ..$ fitted    : Time-Series [1:1265] from 1 to 422: 12.3 12.3 12.5 12.2 12.3 ...
  .. ..$ states    : mts [1:1266, 1:2] 12.4 12.4 12.5 12.3 12.3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.667 422.333 3
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 1 0.44 0.849 12.442 -0.137
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 3
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 12.442 -0.137
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.0368
  .. ..$ x         : Time-Series [1:1265] from 1 to 422: 12.4 12.5 12.3 12.3 12.2 12 11.9 11.9 11.8 11.6 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:6] from 423 to 424: 45.7 45.7 45.6 45.6 45.6 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:1265] from 1 to 422: 12.4 12.5 12.3 12.3 12.2 12 11.9 11.9 11.8 11.6 ...
  ..$ upper    : mts [1:6, 1:2] 45.9 46.1 46.3 46.5 46.6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 423 424 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:6, 1:2] 45.4 45.2 44.9 44.7 44.6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 423 424 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:1265] from 1 to 422: 12.3 12.3 12.5 12.2 12.3 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:1265] from 1 to 422: 0.0748 0.171 -0.2036 0.073 -0.0653 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA223:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -2208
  .. ..$ aic       : num 4427
  .. ..$ bic       : num 4450
  .. ..$ aicc      : num 4427
  .. ..$ mse       : num 1.01
  .. ..$ amse      : num 3.22
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 4417
  .. .. ..$ par    : num [1:5] 0.9999 0.4169 0.8251 14.9133 -0.0283
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 618
  .. ..$ residuals : Time-Series [1:755] from 1 to 755: 0.000676 -0.045969 0.039303 0.001131 -0.047761 ...
  .. ..$ fitted    : Time-Series [1:755] from 1 to 755: 14.9 14.9 14 14.5 14.5 ...
  .. ..$ states    : mts [1:756, 1:2] 14.9 14.9 14.2 14.5 14.5 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0 755 1
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.9999 0.4169 0.8251 14.9133 -0.0283
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 1
  .. ..$ method    : chr "ETS(M,Ad,N)"
  .. ..$ components: chr [1:4] "M" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 14.9133 -0.0283
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.000302
  .. ..$ x         : Time-Series [1:755] from 1 to 755: 14.9 14.2 14.5 14.5 13.8 13.5 13.8 14.2 14.2 13.8 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:10] from 756 to 765: 48.8 48.6 48.6 48.5 48.4 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:755] from 1 to 755: 14.9 14.2 14.5 14.5 13.8 13.5 13.8 14.2 14.2 13.8 ...
  ..$ upper    : mts [1:10, 1:2] 49.8 50.7 51.5 52.1 52.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 756 765 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:10, 1:2] 47.7 46.6 45.7 44.8 44 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 756 765 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:755] from 1 to 755: 14.9 14.9 14 14.5 14.5 ...
  ..$ method   : chr "ETS(M,Ad,N)"
  ..$ residuals: Time-Series [1:755] from 1 to 755: 0.000676 -0.045969 0.039303 0.001131 -0.047761 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA198:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -2046
  .. ..$ aic       : num 4101
  .. ..$ bic       : num 4124
  .. ..$ aicc      : num 4101
  .. ..$ mse       : num 0.41
  .. ..$ amse      : num 1.41
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 4091
  .. .. ..$ par    : num [1:5] 0.977 0.485 0.832 14.414 -0.505
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 491
  .. ..$ residuals : Time-Series [1:763] from 1 to 192: 0.000398 -0.003853 0.000678 0.019529 -0.006452 ...
  .. ..$ fitted    : Time-Series [1:763] from 1 to 192: 14 13.7 13.3 13 13.2 ...
  .. ..$ states    : mts [1:764, 1:2] 14.4 14 13.6 13.3 13.3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.75 191.5 4
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.977 0.485 0.832 14.414 -0.505
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 4
  .. ..$ method    : chr "ETS(M,Ad,N)"
  .. ..$ components: chr [1:4] "M" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 14.414 -0.505
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.000206
  .. ..$ x         : Time-Series [1:763] from 1 to 192: 14 13.6 13.3 13.3 13.1 13.1 13.3 13.3 13.5 13.5 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:8] from 192 to 194: 49 48.9 48.8 48.7 48.7 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:763] from 1 to 192: 14 13.6 13.3 13.3 13.1 13.1 13.3 13.3 13.5 13.5 ...
  ..$ upper    : mts [1:8, 1:2] 49.9 50.7 51.3 52 52.6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 192 194 4
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:8, 1:2] 48.1 47.1 46.3 45.5 44.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 192 194 4
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:763] from 1 to 192: 14 13.7 13.3 13 13.2 ...
  ..$ method   : chr "ETS(M,Ad,N)"
  ..$ residuals: Time-Series [1:763] from 1 to 192: 0.000398 -0.003853 0.000678 0.019529 -0.006452 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA208:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -1591
  .. ..$ aic       : num 3191
  .. ..$ bic       : num 3215
  .. ..$ aicc      : num 3191
  .. ..$ mse       : num 0.0782
  .. ..$ amse      : num 0.198
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 3181
  .. .. ..$ par    : num [1:5] 0.958 0.216 0.962 13.915 -0.13
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 359
  .. ..$ residuals : Time-Series [1:775] from 1 to 259: 0.00987 -0.18169 0.24364 0.30495 -0.15948 ...
  .. ..$ fitted    : Time-Series [1:775] from 1 to 259: 13.8 13.7 13.4 13.5 13.8 ...
  .. ..$ states    : mts [1:776, 1:2] 13.9 13.8 13.5 13.6 13.8 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.667 259 3
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.958 0.216 0.962 13.915 -0.13
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 3
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 13.91 -0.13
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.0782
  .. ..$ x         : Time-Series [1:775] from 1 to 259: 13.8 13.5 13.6 13.8 13.6 13.5 13.1 13.1 12.9 12.2 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:6] from 259 to 261: 45.7 45.5 45.4 45.2 45.1 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:775] from 1 to 259: 13.8 13.5 13.6 13.8 13.6 13.5 13.1 13.1 12.9 12.2 ...
  ..$ upper    : mts [1:6, 1:2] 46.1 46.2 46.2 46.3 46.3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 259 261 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:6, 1:2] 45.4 44.9 44.6 44.2 43.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 259 261 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:775] from 1 to 259: 13.8 13.7 13.4 13.5 13.8 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:775] from 1 to 259: 0.00987 -0.18169 0.24364 0.30495 -0.15948 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA186:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -1344
  .. ..$ aic       : num 2698
  .. ..$ bic       : num 2721
  .. ..$ aicc      : num 2698
  .. ..$ mse       : num 0.04
  .. ..$ amse      : num 0.145
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 2688
  .. .. ..$ par    : num [1:5] 0.854 0.63 0.899 11.447 -0.178
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 379
  .. ..$ residuals : Time-Series [1:781] from 1 to 261: 0.0132 -0.0613 0.1488 -0.1206 0.1025 ...
  .. ..$ fitted    : Time-Series [1:781] from 1 to 261: 11.3 11.2 11 11 10.8 ...
  .. ..$ states    : mts [1:782, 1:2] 11.4 11.3 11.1 11.1 10.9 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.667 261 3
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.854 0.63 0.899 11.447 -0.178
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 3
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 11.447 -0.178
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.04
  .. ..$ x         : Time-Series [1:781] from 1 to 261: 11.3 11.1 11.1 10.9 10.9 10.6 10.6 10.4 10.4 10.4 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:6] from 261 to 263: 44.9 44.8 44.6 44.5 44.4 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:781] from 1 to 261: 11.3 11.1 11.1 10.9 10.9 10.6 10.6 10.4 10.4 10.4 ...
  ..$ upper    : mts [1:6, 1:2] 45.2 45.3 45.5 45.6 45.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 261 263 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:6, 1:2] 44.7 44.2 43.8 43.4 43 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 261 263 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:781] from 1 to 261: 11.3 11.2 11 11 10.8 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:781] from 1 to 261: 0.0132 -0.0613 0.1488 -0.1206 0.1025 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA183:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -512
  .. ..$ aic       : num 1028
  .. ..$ bic       : num 1035
  .. ..$ aicc      : num 1028
  .. ..$ mse       : num 1.2
  .. ..$ amse      : num 2.48
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 1024
  .. .. ..$ par    : num [1:2] 1 16.5
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 99
  .. ..$ residuals : Time-Series [1:189] from 1 to 189: -3.63e-03 -3.67e-07 -3.71e-11 -3.55e-15 0.00 ...
  .. ..$ fitted    : Time-Series [1:189] from 1 to 189: 16.5 16.5 16.5 16.5 16.5 ...
  .. ..$ states    : ts [1:190, 1] 16.5 16.5 16.5 16.5 16.5 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr "l"
  .. .. ..- attr(*, "tsp")= num [1:3] 0 189 1
  .. ..$ par       : Named num [1:2] 1 16.5
  .. .. ..- attr(*, "names")= chr [1:2] "alpha" "l"
  .. ..$ m         : num 1
  .. ..$ method    : chr "ETS(A,N,N)"
  .. ..$ components: chr [1:4] "A" "N" "N" "FALSE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num 16.5
  .. .. ..- attr(*, "names")= chr "l"
  .. ..$ sigma2    : num 1.2
  .. ..$ x         : Time-Series [1:189] from 1 to 189: 16.5 16.5 16.5 16.5 16.5 14.7 14.7 14.7 13.8 13.8 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:10] from 190 to 199: 39.6 39.6 39.6 39.6 39.6 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:189] from 1 to 189: 16.5 16.5 16.5 16.5 16.5 14.7 14.7 14.7 13.8 13.8 ...
  ..$ upper    : mts [1:10, 1:2] 41 41.6 42 42.4 42.7 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 190 199 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:10, 1:2] 38.2 37.6 37.2 36.8 36.5 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 190 199 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:189] from 1 to 189: 16.5 16.5 16.5 16.5 16.5 ...
  ..$ method   : chr "ETS(A,N,N)"
  ..$ residuals: Time-Series [1:189] from 1 to 189: -3.63e-03 -3.67e-07 -3.71e-11 -3.55e-15 0.00 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA231:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -1136
  .. ..$ aic       : num 2281
  .. ..$ bic       : num 2304
  .. ..$ aicc      : num 2281
  .. ..$ mse       : num 0.0256
  .. ..$ amse      : num 0.0595
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 2271
  .. .. ..$ par    : num [1:5] 0.792 0.283 0.962 13.865 -0.388
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 369
  .. ..$ residuals : Time-Series [1:764] from 1 to 382: 0.1092 -0.3472 0.0399 0.1942 0.359 ...
  .. ..$ fitted    : Time-Series [1:764] from 1 to 382: 13.5 13.2 12.6 12.2 12 ...
  .. ..$ states    : mts [1:765, 1:2] 13.9 13.6 13 12.6 12.4 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.5 382.5 2
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.792 0.283 0.962 13.865 -0.388
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 2
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 13.865 -0.388
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.0256
  .. ..$ x         : Time-Series [1:764] from 1 to 382: 13.6 12.9 12.6 12.4 12.4 12.2 12.4 12.4 12.2 11.8 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:4] from 383 to 384: 45.5 45.4 45.2 45.1
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:764] from 1 to 382: 13.6 12.9 12.6 12.4 12.4 12.2 12.4 12.4 12.2 11.8 ...
  ..$ upper    : mts [1:4, 1:2] 45.8 45.7 45.7 45.7 45.9 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 383 384 2
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:4, 1:2] 45.3 45.1 44.8 44.5 45.2 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 383 384 2
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:764] from 1 to 382: 13.5 13.2 12.6 12.2 12 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:764] from 1 to 382: 0.1092 -0.3472 0.0399 0.1942 0.359 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA193:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -1483
  .. ..$ aic       : num 2977
  .. ..$ bic       : num 3000
  .. ..$ aicc      : num 2977
  .. ..$ mse       : num 0.06
  .. ..$ amse      : num 0.173
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 2967
  .. .. ..$ par    : num [1:5] 0.984 0.306 0.931 17.38 -0.206
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 294
  .. ..$ residuals : Time-Series [1:773] from 1 to 258: 0.0121 -0.5244 0.3043 0.8094 -0.027 ...
  .. ..$ fitted    : Time-Series [1:773] from 1 to 258: 17.2 17 16.2 16.3 17.1 ...
  .. ..$ states    : mts [1:774, 1:2] 17.4 17.2 16.5 16.5 17.1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.667 258.333 3
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.984 0.306 0.931 17.38 -0.206
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 3
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 17.38 -0.206
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.06
  .. ..$ x         : Time-Series [1:773] from 1 to 258: 17.2 16.5 16.5 17.1 17.1 17.2 17.2 17.1 16.9 16.7 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:6] from 259 to 260: 44.3 44.2 44.1 44 44 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:773] from 1 to 258: 17.2 16.5 16.5 17.1 17.1 17.2 17.2 17.1 16.9 16.7 ...
  ..$ upper    : mts [1:6, 1:2] 44.6 44.8 44.9 45.1 45.2 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 259 260 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:6, 1:2] 44 43.6 43.3 43 42.7 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 259 260 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:773] from 1 to 258: 17.2 17 16.2 16.3 17.1 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:773] from 1 to 258: 0.0121 -0.5244 0.3043 0.8094 -0.027 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA167:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -1256
  .. ..$ aic       : num 2522
  .. ..$ bic       : num 2545
  .. ..$ aicc      : num 2522
  .. ..$ mse       : num 0.0458
  .. ..$ amse      : num 0.124
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 2512
  .. .. ..$ par    : num [1:5] 0.965 0.266 0.946 16.106 -0.117
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 400
  .. ..$ residuals : Time-Series [1:719] from 1 to 719: 0.00432 -0.09645 0.11883 -0.11018 -0.09495 ...
  .. ..$ fitted    : Time-Series [1:719] from 1 to 719: 16 15.9 15.7 15.7 15.5 ...
  .. ..$ states    : mts [1:720, 1:2] 16.1 16 15.8 15.8 15.6 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0 719 1
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.965 0.266 0.946 16.106 -0.117
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 1
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 16.106 -0.117
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.0458
  .. ..$ x         : Time-Series [1:719] from 1 to 719: 16 15.8 15.8 15.6 15.4 15.4 15.6 16.2 16.3 16 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:10] from 720 to 729: 47.6 47.5 47.4 47.3 47.2 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:719] from 1 to 719: 16 15.8 15.8 15.6 15.4 15.4 15.6 16.2 16.3 16 ...
  ..$ upper    : mts [1:10, 1:2] 47.9 48 48.1 48.1 48.2 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 720 729 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:10, 1:2] 47.3 47 46.7 46.4 46.2 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 720 729 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:719] from 1 to 719: 16 15.9 15.7 15.7 15.5 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:719] from 1 to 719: 0.00432 -0.09645 0.11883 -0.11018 -0.09495 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA132:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -993
  .. ..$ aic       : num 1997
  .. ..$ bic       : num 2017
  .. ..$ aicc      : num 1997
  .. ..$ mse       : num 0.161
  .. ..$ amse      : num 0.474
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 1987
  .. .. ..$ par    : num [1:5] 0.998 0.31 0.941 9.631 -0.245
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 251
  .. ..$ residuals : Time-Series [1:461] from 1 to 154: -0.000307 0.116958 -0.229782 0.026804 -0.093968 ...
  .. ..$ fitted    : Time-Series [1:461] from 1 to 154: 9.4 9.18 9.13 8.67 8.49 ...
  .. ..$ states    : mts [1:462, 1:2] 9.63 9.4 9.3 8.9 8.7 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0.667 154.333 3
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.998 0.31 0.941 9.631 -0.245
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 3
  .. ..$ method    : chr "ETS(A,Ad,N)"
  .. ..$ components: chr [1:4] "A" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 9.631 -0.245
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.161
  .. ..$ x         : Time-Series [1:461] from 1 to 154: 9.4 9.3 8.9 8.7 8.4 8.4 8.3 8.2 8.2 7.9 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:6] from 155 to 156: 45 44.5 44.1 43.7 43.3 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:461] from 1 to 154: 9.4 9.3 8.9 8.7 8.4 8.4 8.3 8.2 8.2 7.9 ...
  ..$ upper    : mts [1:6, 1:2] 45.5 45.5 45.4 45.4 45.4 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 155 156 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:6, 1:2] 44.5 43.6 42.7 42 41.2 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 155 156 3
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:461] from 1 to 154: 9.4 9.18 9.13 8.67 8.49 ...
  ..$ method   : chr "ETS(A,Ad,N)"
  ..$ residuals: Time-Series [1:461] from 1 to 154: -0.000307 0.116958 -0.229782 0.026804 -0.093968 ...
  ..- attr(*, "class")= chr "forecast"
 $ KILCHICA174:List of 9
  ..$ model    :List of 18
  .. ..$ loglik    : num -2290
  .. ..$ aic       : num 4591
  .. ..$ bic       : num 4614
  .. ..$ aicc      : num 4591
  .. ..$ mse       : num 0.646
  .. ..$ amse      : num 2.17
  .. ..$ fit       :List of 4
  .. .. ..$ value  : num 4581
  .. .. ..$ par    : num [1:5] 0.962 0.542 0.8 17.779 -0.367
  .. .. ..$ fail   : int 0
  .. .. ..$ fncount: int 284
  .. ..$ residuals : Time-Series [1:790] from 1 to 790: 0.00655 -0.00058 0.00884 0.00357 0.00117 ...
  .. ..$ fitted    : Time-Series [1:790] from 1 to 790: 17.5 17.4 17.2 17.3 17.4 ...
  .. ..$ states    : mts [1:791, 1:2] 17.8 17.6 17.4 17.4 17.4 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "l" "b"
  .. .. ..- attr(*, "tsp")= num [1:3] 0 790 1
  .. .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  .. ..$ par       : Named num [1:5] 0.962 0.542 0.8 17.779 -0.367
  .. .. ..- attr(*, "names")= chr [1:5] "alpha" "beta" "phi" "l" ...
  .. ..$ m         : num 1
  .. ..$ method    : chr "ETS(M,Ad,N)"
  .. ..$ components: chr [1:4] "M" "A" "N" "TRUE"
  .. ..$ call      : language ets(y = object, lambda = lambda, allow.multiplicative.trend = allow.multiplicative.trend)
  .. ..$ initstate : Named num [1:2] 17.779 -0.367
  .. .. ..- attr(*, "names")= chr [1:2] "l" "b"
  .. ..$ sigma2    : num 0.000282
  .. ..$ x         : Time-Series [1:790] from 1 to 790: 17.6 17.4 17.4 17.4 17.4 17.1 16.9 16.9 16.7 16.7 ...
  .. ..- attr(*, "class")= chr "ets"
  ..$ mean     : Time-Series [1:10] from 791 to 800: 47.3 47.2 47.1 47 46.9 ...
  ..$ level    : num [1:2] 80 95
  ..$ x        : Time-Series [1:790] from 1 to 790: 17.6 17.4 17.4 17.4 17.4 17.1 16.9 16.9 16.7 16.7 ...
  ..$ upper    : mts [1:10, 1:2] 48.3 49.2 50 50.7 51.3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 791 800 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ lower    : mts [1:10, 1:2] 46.3 45.1 44.2 43.3 42.5 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Series 1" "Series 2"
  .. ..- attr(*, "tsp")= num [1:3] 791 800 1
  .. ..- attr(*, "class")= chr [1:3] "mts" "ts" "matrix"
  ..$ fitted   : Time-Series [1:790] from 1 to 790: 17.5 17.4 17.2 17.3 17.4 ...
  ..$ method   : chr "ETS(M,Ad,N)"
  ..$ residuals: Time-Series [1:790] from 1 to 790: 0.00655 -0.00058 0.00884 0.00357 0.00117 ...
  ..- attr(*, "class")= chr "forecast"
}}
\usage{
Chicago.forecast.10day
}
\description{
This contains at 10-day Forecast over all the PWS using wundr's kriging method
}
\examples{
data(Chicago.forecast.10day)
}
\author{
wundr team
}
\keyword{datasets}

