% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{create_geo_cond}
\alias{create_geo_cond}
\title{create_geo_cond}
\usage{
create_geo_cond(data.conditions, variable)
}
\arguments{
\item{data.conditions}{Data frame with data on conditions for various weather stations}

\item{variable}{Name of variable to select from data frame}
}
\value{
A object of type geodata containing coordinates (lon,lat) and values for selected variable
}
\description{
Takes as input a data frame of conditions from weather stations and selects coordinates as well as
conditions for one variable and saves them as a geodata object
}
\examples{
data(Rio_conditions)
data.geo <- create_geo_cond(Rio_conditions,"temp_c")
data.geo

}

