% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{create_grid}
\alias{create_grid}
\title{create_grid}
\usage{
create_grid(data.geo, size.lon = 50, size.lat = 50, grid.lim = NULL)
}
\arguments{
\item{data.geo}{geodata object with coordinates in attribute called 'coords'}

\item{size.lon}{Discrte size (number of grid points) in longitudinal direction (defaultis 50).}

\item{size.lat}{Discrte size (number of grid points) in latitudinal direction (default is 50).}

\item{grid.lim}{Limiting values of longitude and latitude for grid. When set to null, it chooses the max/min value of longitude and latitude in coord, otherwise a vector c(min.lon,max.lon,min.lat,max.lat) must be provided}
}
\value{
A data frame containing the grid points with columns labelled (lon,lat)
}
\description{
Creates a grid of points in a rectangular over a sets of points coordinates from a geodata object
or any other object with coordinates saves in an attribute called 'coords'
}
\examples{
data <- Rio_conditions
data.geo <- create_geo_cond(Rio_conditions,"temp_c")
grid.positions <- create_grid(data.geo)
grid.positions <- create_grid(data.geo, grid.lim= c(-43.6,-42.8,-23.2,-22.6) )
head(grid.positions)

}

