% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\name{createCentroidTable}
\alias{createCentroidTable}
\title{createCentroidTable}
\usage{
createCentroidTable(longitude, latitude, radius, max_radius_km)
}
\arguments{
\item{longitude}{Standard Longitude in range -/+180}

\item{latitude}{Standard latitude in range -/+90}

\item{radius}{Radius in kilometers of the large circle to be covered by small cricles}

\item{max_radius_km}{Radius in kilometers of the small cricles (e.g. 40km)}
}
\value{
A data frame of the (longitude,latitude) pairs for the centres of the small circles which cover the large circle
}
\description{
createCentroidTable is an auxiliary function used in the 'PWS_meta_query' function.
Given a the coordinates of the centre of a large circle and its radius,
the createCentroidTable function creates a data frame of the (longitude,latitude)-pairs
for the centres of the small circles of a given radius which cover the large circle.
}
\examples{
centre_table <- createCentroidTable(-122,37,400,40)
plot(centre_table,pch=18)
}

