% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetting_functions.R
\name{draw_subset}
\alias{draw_subset}
\title{draw_subset}
\usage{
draw_subset(PWS.class)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}
}
\value{
A PWS.class points S4 object.
}
\description{
Graphicly subset PWS.class objects by drawing a poylgon around PWS points
plotted in graphics device. Adapted from Roger Bivand's Applied Spatial Data
Analysis with R (pp. 76-78).
}
\examples{
\dontrun{
# not run because this is an function that requires user interaction.
my_ss  <-  draw_subset(PWS.Conds.Chicago)
my_ss2 <-  draw_subset(my_ss)}
}

