% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartoDB_functions.R
\name{get_cdb_table}
\alias{get_cdb_table}
\title{get_cdb_table}
\usage{
get_cdb_table(table_name, cdb_account)
}
\arguments{
\item{table_name}{A PostGres-based table from CartoDB.}

\item{cdb_account}{A CartoDB account name.}
}
\value{
A data frame.
}
\description{
... yada yada
}
\examples{
# test CartoDB connection by retrieving existing table ...
stations <- get_cdb_table("public.stations", your.cdb.account)
}

