% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{gg_points}
\alias{gg_points}
\title{gg_points}
\usage{
gg_points(PWS.class, basemap = basemap, title = NULL, ...)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{basemap}{A contextual basemap. See set_basemap.}

\item{title}{A character string title for the map. Default = NULL.}

\item{...}{...}
}
\description{
Plots PWS locations on a contextual basemap.
}
\examples{
data("PWS.Conds.Chicago")
basemap <- set_basemap(PWS.Conds.Chicago, zoom = 12)
gg_points(PWS.Conds.Chicago, basemap, title = "Downtown Chicago PWS")
}

