% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{history_forecast}
\alias{history_forecast}
\title{history_forecast}
\usage{
history_forecast(history.tszoo, find.frequency = TRUE, ...)
}
\arguments{
\item{history.tszoo}{Time series object from history_ts or history_zoo}

\item{find.frequency}{Boolean. Whether to find frequency automatically (default is TURE).}

\item{...}{Additional parameters to be passed to 'forecast'}
}
\value{
Time series object of type 'ts' with selected variables measured by selected weather station
}
\description{
Forecasts of time series created from 'history_zoo' or 'history_ts'. Note that the time series must be univariate.
}
\examples{
data(Rio_history)
hist.ts <- history_ts(Rio_history,"IRIODEJA53","hum")
hist.forecast <- history_forecast(hist.ts)
plot(hist.forecast, main = 'Forecast', xlab='Time (days)', ylab='Humidity (\%)')

}

