% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{history_ts}
\alias{history_ts}
\title{history_ts}
\usage{
history_ts(hist.data, id, variables)
}
\arguments{
\item{hist.data}{data frame of historical weather data (e.g. as created by 'PWS_history')}

\item{id}{Personal Weather Station ID}

\item{variables}{Variable(s) to be selected as character (can be a vector, see example below)}
}
\value{
Time series object of type 'ts' with selected variables measured by selected weather station
}
\description{
Regular times series object of historical weather data. Takes as input a data frame of historical
weather data e.g. as created by 'PWS_history' or corresponding S4 function. It selects the data for a
given station and given variable(s) and returns a time series object of class 'ts'.
}
\examples{
data(Rio_history)
hist.ts <- history_ts(Rio_history,"IRIODEJA53","hum")
plot(hist.ts,col='red', main = "Humidity")

}

