% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{history_zoo}
\alias{history_zoo}
\title{history_zoo}
\usage{
history_zoo(hist.data, id, variables)
}
\arguments{
\item{hist.data}{data frame of historical weather data (e.g. as created by 'PWS_history')}

\item{id}{Personal Weather Station ID}

\item{variables}{Variable(s) to be selected as character (can be a vector, see example below)}
}
\value{
Time series object of type 'zoo' with selected variables measured by selected weather station
}
\description{
Irregular times series object of historical weather data. Takes as input a data frame of historical
weather data e.g. as created by 'PWS_history' or corresponding S4 function. It selects the data for a
given station and given variable(s) and returns a time series object of class 'zoo' (which is designed
for irregular time series).
}
\examples{
data(Rio_history)
hist.zoo <- history_zoo(Rio_history,"IRIODEJA53",c("hum","tempm"))
plot(hist.zoo,col='red', main = "Humidity and Temperatur")
}

