% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\docType{methods}
\name{initialize,PWS.Conditions-method}
\alias{initialize,PWS.Conditions-method}
\title{S4 PWS.Conditions Initializer function}
\usage{
\S4method{initialize}{PWS.Conditions}(.Object, PWS.Locations, user.key, ...)
}
\arguments{
\item{.Object}{S4 initializer object}

\item{PWS.Locations}{A PWS.Locations S4 Class Object}

\item{user.key}{character vector of user.key}

\item{...}{other user input}
}
\description{
S4 PWS.Conditions Initializer function
}

