<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
% Please edit documentation in R/S4.History.R
\docType{methods}
\name{initialize,PWS.History-method}
\alias{initialize,PWS.History-method}
\title{S4 Initializer function}
\usage{
\S4method{initialize}{PWS.History}(.Object, PWS.Locations, begin_YYYYMMDD,
  end_YYYYMMDD, user.key, imperial = TRUE, ...)
}
\arguments{
\item{.Object}{S4 initializer object}

\item{PWS.Locations}{S4 Object Required}

\item{begin_YYYYMMDD}{begin of query}

\item{end_YYYYMMDD}{end of query}

\item{user.key}{user.key of user}

\item{imperial}{whether the user wants imperial or metric}

<<<<<<< HEAD
\item{...}{sundry}
=======
\item{...}{Passed on to next function}
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
}
\description{
S4 Initializer function
}

