% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Locations.R
\docType{methods}
\name{initialize,PWS.Locations-method}
\alias{initialize,PWS.Locations-method}
\title{S4 Initializer function}
\usage{
\S4method{initialize}{PWS.Locations}(.Object, longitude, latitude, radius,
  user.key)
}
\arguments{
\item{.Object}{S4 initializer object}

\item{longitude}{desired longitude}

\item{latitude}{desired latitude}

\item{radius}{desired radius}

\item{user.key}{user.key of user}
}
\description{
S4 Initializer function
}

