% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Query.Subset.R
\docType{methods}
\name{initialize,PWS.Query.Subset-method}
\alias{initialize,PWS.Query.Subset-method}
\title{S4 PWS.Query.Subset Initializer function}
\usage{
\S4method{initialize}{PWS.Query.Subset}(.Object, PWS.Locations, longitude,
  latitude, radius, ...)
}
\arguments{
\item{.Object}{S4 initializer object}

\item{PWS.Locations}{A PWS.Locations S4 Class Object}

\item{longitude}{User's Longitude of Choice}

\item{latitude}{User's Latitude of Choice}

\item{radius}{desired within confines of larger search}

\item{...}{other user input}
}
\description{
S4 PWS.Query.Subset Initializer function
}

