% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Locations.R
\name{j.geocode}
\alias{j.geocode}
\title{j.geocode}
\usage{
j.geocode(address)
}
\arguments{
\item{address}{name}
}
\value{
Coordinates of desired location
}
\description{
j.geocode
}
\examples{
j.geocode("Santa Monica, CA")
}

