% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\docType{class}
\name{PWS.Conditions-class}
\alias{PWS.Conditions-class}
\title{An S4 class to represent the PWS Conditions of a user search}
\description{
An S4 class to represent the PWS Conditions of a user search
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{A SpatialPointsDataFrame}

\item{\code{spatialPt}}{A SpatialPoints Matrix}

\item{\code{call}}{list of user's desired location coordinates and radius}

\item{\code{data}}{data}
}}

