% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\name{PWS.Conditions}
\alias{PWS.Conditions}
\title{PWS.Conditions constructor function}
\usage{
PWS.Conditions(...)
}
\arguments{
\item{...}{coordinates of PWS.Conditions initializer function}
}
\value{
S4 PWS.Condition class
}
\description{
PWS.Conditions constructor function
}

