% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\docType{data}
\name{PWS.Conds.Chicago}
\alias{PWS.Conds.Chicago}
\title{Chicago Conditions Dataset}
\format{An object of class \code{PWS.Conditions} of length 1.}
\usage{
PWS.Conds.Chicago
}
\description{
This contains the weather conditions for the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
called with PWS.Conds.Chicago <- PWS.Conditions(PWS.Loc.Chicago, user.key=jam.key.2)
where the PWS.Loc.Chicago call is described in the PWS.Locations S4 Class
}
\examples{
data(PWS.Conds.Chicago)
}
\author{
wundr team
}
\keyword{datasets}

