% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.History.R
\docType{data}
\name{PWS.Hist.Chicago}
\alias{PWS.Hist.Chicago}
\title{Chicago History Dataset}
\format{An object of class \code{PWS.History} of length 1.}
\usage{
PWS.Hist.Chicago
}
\description{
This contains the weather history for the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
called with PWS.Hist.Chicago <- PWS.History(PWS.Loc.Chicago, "20150306", "20150310", user.key)
where the PWS.Loc.Chicago call is described in the PWS.Locations S4 Class
}
\examples{
data(PWS.Hist.Chicago)
}
\author{
wundr team
}
\keyword{datasets}

