% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.History.R
\docType{class}
\name{PWS.History-class}
\alias{PWS.History-class}
\title{PWS.History
An S4 class to store history data related to PWS searches}
\value{
S4 class
}
\description{
PWS.History
An S4 class to store history data related to PWS searches
}
\section{Slots}{

\describe{
\item{\code{average.Temp}}{values}

\item{\code{average.Humidity}}{values}

\item{\code{average.Pressure}}{values}

\item{\code{average.Dew.Point}}{values}

\item{\code{variance.Temp}}{values}

\item{\code{variance.Humidity}}{values}

\item{\code{variance.Pressure}}{values}

\item{\code{variance.Dew.Point}}{values}

\item{\code{standard.Dev.Temp}}{values}

\item{\code{standard.Dev.Humidity}}{values}

\item{\code{standard.Dev.Pressure}}{values}

\item{\code{standard.Dev.Dew.Point}}{values}

\item{\code{spatialPtDF}}{ANY SpatialPointsDataFrame}

\item{\code{spatialPt}}{ANY SpatialPoints}

\item{\code{call}}{list of query search}

\item{\code{history}}{data}
}}

