<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.History.R
\name{PWS.History}
\alias{PWS.History}
\title{PWS.Locations constructor function}
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.History.R
\name{PWS.History}
\alias{PWS.History}
\title{PWS.History constructor function}
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
\usage{
PWS.History(...)
}
\arguments{
\item{...}{coordinates of initializer function}
}
\value{
S4 class
}
\description{
<<<<<<< HEAD
PWS.Locations constructor function
=======
PWS.History constructor function
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
}

