<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
% Please edit documentation in R/S4.Locations.R
\docType{data}
\name{PWS.Loc.Chicago}
\alias{PWS.Loc.Chicago}
\title{Chicago Locations Dataset}
<<<<<<< HEAD
\format{An object of class \code{PWS.Locations} of length 1.}
=======
\format{\preformatted{Formal class 'PWS.Locations' [package "wundr"] with 3 slots
  ..@ spatialPtDF:Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  .. .. ..@ data       :'data.frame':	24 obs. of  9 variables:
  .. .. .. ..$ neighborhood: chr [1:24] "South Loop (Central Station)" "South Loop" "Near South Side" "Near South Side" ...
  .. .. .. ..$ city        : chr [1:24] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. .. .. ..$ state       : chr [1:24] "IL" "IL" "IL" "IL" ...
  .. .. .. ..$ country     : chr [1:24] "US" "US" "US" "US" ...
  .. .. .. ..$ id          : chr [1:24] "KILCHICA310" "KILCHICA388" "KILCHICA315" "KILCHICA166" ...
  .. .. .. ..$ lat         : num [1:24] 41.9 41.9 41.9 41.9 41.9 ...
  .. .. .. ..$ lon         : num [1:24] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..$ distance_km : num [1:24] 1.61 1.99 1.95 2.38 1.03 ...
  .. .. .. ..$ distance_mi : num [1:24] 1.003 1.234 1.213 1.476 0.643 ...
  .. .. ..@ coords.nrs : num(0) 
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ spatialPt  :Formal class 'SpatialPoints' [package "sp"] with 3 slots
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ call       :List of 3
  .. ..$ lon      : num -87.6
  .. ..$ lat      : num 41.9
  .. ..$ radius_km: num 5
}}
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
\usage{
PWS.Loc.Chicago
}
\description{
This contains a listing of the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
(PWS.Loc.Chicago <- PWS.Locations("Chicago, IL", radius=5, user.key)).
}
\examples{
data(PWS.Loc.Chicago)
<<<<<<< HEAD
head(PWS.Loc.Chicago)
=======
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
}
\author{
wundr team
}
\keyword{datasets}

