% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Locations.R
\docType{class}
\name{PWS.Locations-class}
\alias{PWS.Locations-class}
\title{PWS.Locations
An S4 class to store and display data related to PWS searches}
\value{
S4 class
}
\description{
PWS.Locations
An S4 class to store and display data related to PWS searches
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{ANY SpatialPointsDataFrame}

\item{\code{spatialPt}}{ANY SpatialPoints}

\item{\code{call}}{list of called coordinates}
}}

