% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Locations.R
\name{PWS.Locations}
\alias{PWS.Locations}
\title{PWS.Locations constructor function}
\usage{
PWS.Locations(...)
}
\arguments{
\item{...}{coordinates of initializer function}
}
\value{
S4 class
}
\description{
PWS.Locations constructor function
}

