% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Query.Subset.R
\docType{class}
\name{PWS.Query.Subset-class}
\alias{PWS.Query.Subset-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{A SpatialPointsDataFrame}

\item{\code{spatialPt}}{A SpatialPoints}

\item{\code{call}}{list of stuff}
}}

