% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Query.Subset.R
\name{PWS.Query.Subset}
\alias{PWS.Query.Subset}
\title{PWS.Query.Subset constructor function}
\usage{
PWS.Query.Subset(...)
}
\arguments{
\item{...}{Additional Details Specified by User}
}
\value{
S4 PWS.Query.Subset class
}
\description{
PWS.Query.Subset constructor function
}

