% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{PWS_conditions}
\alias{PWS_conditions}
\title{PWS_conditions}
\usage{
PWS_conditions(PWSmetadata, user_key, stdAPI = TRUE)
}
\arguments{
\item{PWSmetadata}{Meta data object of weather stations (output of 'PWS_meta_query')}

\item{user_key}{Your WUnderground API user key given as a character string}

\item{stdAPI}{A bolean variable indicating whether you have a standard (free) API access (TRUE) which only allows for 10 API calls per minute. The default is the standard API.}
}
\value{
A data frame containing all the weather conditions of the stations in the provided meta data object.
}
\description{
Given an output of 'PWS_meta_query' which contains meta data of personal weather stations in a region
the function 'PWS_conditions' downloads the weather conditions for all stations in the meta data.
}
\examples{
#' # your.key <- "xxxxxxxxxx" # replace this with your key
# cond <- PWS_conditions(pwsmetadata2,stefan.key)
# if you run the above code with your key the output should be the same as provided here:
# TODO ++++ SAVE DATA AND LOAD IT HERE
# head(cond)

}

