% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\name{PWS_history}
\alias{PWS_history}
\title{PWS_history}
\usage{
PWS_history(PWSmetadata, begin_YYYYMMDD, end_YYYYMMDD, user_key,
  stdAPI = TRUE)
}
\arguments{
\item{PWSmetadata}{Meta data object of weather stations (output of 'PWS_meta_query')}

\item{begin_YYYYMMDD}{Beginning date in format "YYYYMMDD"}

\item{end_YYYYMMDD}{Ending date in format "YYYYMMDD"}

\item{user_key}{Your WUnderground API user key given as a character string}

\item{stdAPI}{A bolean variable indicating whether you have a standard (free) API access (TRUE) which only allows for 10 API calls per minute. The default is the standard API.}
}
\value{
A data frame containing all the weather conditions of the stations in the provided meta data object.
}
\description{
Given an output of 'PWS_meta_query' which contains meta data of personal weather stations in a region
and a start and an end date for a period of time, the function 'PWS_history' downloads the weather
conditions history for all stations in the meta data. Note that 1 API per stations per day is required.
}
\examples{
# your.key <- "xxxxxxxxxx" # replace this with your key
# hist <- PWS_history(pwsmetadata,"20160101","20160101",your.key)
# if you run the above code with your key the output should be the same as provided here:
# TODO ++++ SAVE DATA AND LOAD IT HERE
# head(hist)
}

