% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{PWS_meta_subset}
\alias{PWS_meta_subset}
\title{PWS_meta_subset}
\usage{
PWS_meta_subset(PWSmetadata, longitude, latitude, radius, km_miles = TRUE)
}
\arguments{
\item{PWSmetadata}{Meta data object of weather stations (output of 'PWS_meta_query')}

\item{longitude}{Standard Longitude in range -/+180}

\item{latitude}{Standard latitude in range -/+90}

\item{radius}{Radius in kilometers of the circle}

\item{km_miles}{A bolean variable indicating whether radius is in kilometer (TRUE) or miles (FALSE). The default is kilometers.}
}
\value{
A list of two entries. The first has information on the original arguments of the function (radius and coordinates of circle). The second entry contains a subsetted data frame with all personal weather stations in the region of the circle.
}
\description{
Given an output of 'PWS_meta_query' which contains meta data of personal weather stations in a region
and the coordinates of the centre of a (new) circle and its radius, the 'PWS_meta_subset' function
subsets the meta data to only cotain stations inside the new circle. The output is the subsetted meta data.
}
\examples{
# First we load the output of the command
# Rio_metadata <- PWS_meta_query(-22.856878, -43.185368, 50, your.key)
# where our.key is your API-key for Weather Underground.
data(Rio_metadata)
# Now we subset:
Rio_centre_metadata <- PWS_meta_subset(Rio_metadata,-43.185368,-22.856878, 10)
# and plot it
data(Rio_basemap)
ggmap(Rio_basemap)+geom_point(data=Rio_centre_metadata$PWSmetadata,col='red')

}

