% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resetOptions}
\alias{resetOptions}
\title{Reset the global package variable \code{wundr.options}}
\usage{
resetOptions()
}
\value{
the default value value of wundr package global \code{wundr.options}
}
\description{
Reset the global package variable \code{wundr.options}
}
\examples{
\dontrun{
  resetOptions()
}

}

