% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_basemap}
\alias{Rio_basemap}
\title{Rio_basemap object}
\format{An object of class \code{ggmap} (inherits from \code{raster}) with 639 rows and 640 columns.}
\usage{
Rio_basemap
}
\description{
Object of tyoe ggmapraster containing the base map of Rio de Janiero. For visualisation
of example data sets. The base map can be plotted using 'ggmap'
}
\examples{
data(Rio_basemap)
ggmap::ggmap(Rio_basemap)

}
\author{
retrived from google maps
}
\keyword{datasets}

