% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_conditions}
\alias{Rio_conditions}
\title{Rio_conditions dataset}
\format{\preformatted{'data.frame':	34 obs. of  70 variables:
 $ full                            : chr  "Galeao-Antonio C Jobim International, Brazil" "Rio De Janeiro, Brazil" "Santos Dumont, Brazil" "Niteroi, Brazil" ...
 $ city                            : chr  "Galeao-Antonio C Jobim International" "Rio De Janeiro" "Santos Dumont" "Niteroi" ...
 $ state                           : chr  "" "" "" "" ...
 $ state_name                      : chr  "Brazil" "Brazil" "Brazil" "Brazil" ...
 $ country                         : chr  "BZ" "BZ" "BZ" "BZ" ...
 $ country_iso3166                 : chr  "BR" "BR" "BR" "BR" ...
 $ zip                             : num  0 0 0 0 0 0 0 0 0 0 ...
 $ magic                           : num  3 1 3 6 3 1 1 1 3 1 ...
 $ wmo                             : chr  "83746" "83755" "83755" "83755" ...
 $ latitude                        : num  -22.8 -22.9 -22.9 -22.9 -22.9 ...
 $ longitude                       : num  -43.2 -43.2 -43.2 -43.1 -43.2 ...
 $ elevation                       : num  8 3 3 15 3 6 3 6 3 3 ...
 $ display_location.full           : chr  "Galeao-Antonio C Jobim International, Brazil" "Rio De Janeiro, Brazil" "Santos Dumont, Brazil" "Niteroi, Brazil" ...
 $ display_location.city           : chr  "Galeao-Antonio C Jobim International" "Rio De Janeiro" "Santos Dumont" "Niteroi" ...
 $ display_location.state          : chr  "" "" "" "" ...
 $ display_location.state_name     : chr  "Brazil" "Brazil" "Brazil" "Brazil" ...
 $ display_location.country        : chr  "BZ" "BZ" "BZ" "BZ" ...
 $ display_location.country_iso3166: chr  "BR" "BR" "BR" "BR" ...
 $ display_location.zip            : num  0 0 0 0 0 0 0 0 0 0 ...
 $ display_location.magic          : num  3 1 3 6 3 1 1 1 3 1 ...
 $ display_location.wmo            : chr  "83746" "83755" "83755" "83755" ...
 $ display_location.latitude       : num  -22.8 -22.9 -22.9 -22.9 -22.9 ...
 $ display_location.longitude      : num  -43.2 -43.2 -43.2 -43.1 -43.2 ...
 $ display_location.elevation      : num  8 3 3 15 3 6 3 6 3 3 ...
 $ station_id                      : chr  "IRIODEJA53" "IRIODEJA63" "IRIODEJA79" "INITERI3" ...
 $ observation_time                : chr  "Last Updated on March 12, 2:42 PM BRT" "Last Updated on March 12, 2:42 PM BRT" "Last Updated on March 12, 2:41 PM BRT" "Last Updated on March 12, 2:42 PM BRT" ...
 $ observation_time_rfc822         : chr  "Sat, 12 Mar 2016 14:42:29 -0300" "Sat, 12 Mar 2016 14:42:49 -0300" "Sat, 12 Mar 2016 14:41:44 -0300" "Sat, 12 Mar 2016 14:42:44 -0300" ...
 $ observation_epoch               : num  1.46e+09 1.46e+09 1.46e+09 1.46e+09 1.46e+09 ...
 $ local_time_rfc822               : chr  "Sat, 12 Mar 2016 14:43:22 -0300" "Sat, 12 Mar 2016 14:43:22 -0300" "Sat, 12 Mar 2016 14:43:23 -0300" "Sat, 12 Mar 2016 14:43:23 -0300" ...
 $ local_epoch                     : num  1.46e+09 1.46e+09 1.46e+09 1.46e+09 1.46e+09 ...
 $ local_tz_short                  : chr  "BRT" "BRT" "BRT" "BRT" ...
 $ local_tz_long                   : chr  "America/Sao_Paulo" "America/Sao_Paulo" "America/Sao_Paulo" "America/Sao_Paulo" ...
 $ local_tz_offset                 : num  -300 -300 -300 -300 -300 -300 -300 -300 -300 -300 ...
 $ weather                         : chr  "Scattered Clouds" "Mostly Cloudy" "Mostly Cloudy" "Mostly Cloudy" ...
 $ temperature_string              : chr  "87.4 F (30.8 C)" "85.8 F (29.9 C)" "83.8 F (28.8 C)" "87.3 F (30.7 C)" ...
 $ temp_f                          : num  87.4 85.8 83.8 87.3 84.9 85.6 84.5 85.8 79.9 82.2 ...
 $ temp_c                          : num  30.8 29.9 28.8 30.7 29.4 29.8 29.2 29.9 26.6 27.9 ...
 $ relative_humidity               : chr  "69\%" "68\%" "69\%" "66\%" ...
 $ wind_string                     : chr  "Calm" "Calm" "From the SSW at 25.3 MPH Gusting to 25.3 MPH" "Calm" ...
 $ wind_dir                        : chr  "North" "North" "SSW" "North" ...
 $ wind_degrees                    : num  -9999 -9999 202 -9999 -9999 ...
 $ wind_mph                        : num  -9999 -9999 25.3 -9999 -9999 ...
 $ wind_gust_mph                   : num  0 0 25.3 0 0 15 12 11.4 23 0 ...
 $ wind_kph                        : num  0 0 40.7 0 0 14.3 9.7 5 19.3 0 ...
 $ wind_gust_kph                   : num  0 0 40.7 0 0 24.1 19.3 18.3 37 0 ...
 $ pressure_mb                     : num  1015 1012 1012 1012 1012 ...
 $ pressure_in                     : num  30 29.9 29.9 29.9 29.9 ...
 $ pressure_trend                  : chr  "-" "0" "0" "0" ...
 $ dewpoint_string                 : chr  "76 F (25 C)" "74 F (24 C)" "72 F (22 C)" "75 F (24 C)" ...
 $ dewpoint_f                      : num  76 74 72 75 74 72 75 70 74 75 ...
 $ dewpoint_c                      : num  25 24 22 24 24 22 24 21 24 24 ...
 $ heat_index_string               : chr  "98 F (37 C)" "94 F (34 C)" "90 F (32 C)" "96 F (36 C)" ...
 $ heat_index_f                    : chr  "98" "94" "90" "96" ...
 $ heat_index_c                    : chr  "37" "34" "32" "36" ...
 $ windchill_string                : chr  "NA" "NA" "NA" "NA" ...
 $ windchill_f                     : chr  "NA" "NA" "NA" "NA" ...
 $ windchill_c                     : chr  "NA" "NA" "NA" "NA" ...
 $ feelslike_string                : chr  "98 F (37 C)" "94 F (34 C)" "90 F (32 C)" "96 F (36 C)" ...
 $ feelslike_f                     : num  98 94 90 96 92 91 93 90 84 89 ...
 $ feelslike_c                     : num  37 34 32 36 34 33 34 32 29 32 ...
 $ visibility_mi                   : num  6.2 6.2 6.2 6.2 6.2 6.2 6.2 6.2 6.2 6.2 ...
 $ visibility_km                   : num  10 10 10 10 10 10 10 10 10 10 ...
 $ solarradiation                  : chr  "--" "--" "--" "--" ...
 $ UV                              : chr  "8" "5" "5" "5" ...
 $ precip_1hr_string               : chr  "-999.00 in ( 0 mm)" "-999.00 in ( 0 mm)" "0.00 in ( 0 mm)" "-999.00 in ( 0 mm)" ...
 $ precip_1hr_in                   : num  -999 -999 0 -999 -999 0 0 0 0 -999 ...
 $ precip_1hr_metric               : chr  " 0" " 0" " 0" " 0" ...
 $ precip_today_string             : chr  "-999.00 in (-25375 mm)" "-999.00 in (-25375 mm)" "1.31 in (33 mm)" "-999.00 in (-25375 mm)" ...
 $ precip_today_in                 : chr  "-999.00" "-999.00" "1.31" "-999.00" ...
 $ precip_today_metric             : chr  "--" "--" "33" "--" ...
}}
\usage{
Rio_conditions
}
\description{
This contains the weather conditions for the Personal Weather Stations in
Rio de Janeiro, Brazil, downloaded on March 12, 2016.
It is the output of the following call of the 'PWS_conditions' function,
Rio_metadata <- PWS_meta_query(-22.856878, -43.185368, 50, your.key)
where, your.key' should be replaced with your API-key for Weather Underground
and 'Rio_metadata' is provided in the package (load it via data(Rio_metadata)).
}
\examples{
data(Rio_conditions)
head(Rio_conditions)
}
\author{
wundr team
}
\keyword{datasets}

