% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_history}
\alias{Rio_history}
\title{Rio_history dataset}
\format{\preformatted{'data.frame':	40202 obs. of  34 variables:
 $ pretty        : chr  "12:09 AM BRST on December 24, 2015" "12:19 AM BRST on December 24, 2015" "12:29 AM BRST on December 24, 2015" "12:46 AM BRST on December 24, 2015" ...
 $ year          : num  2015 2015 2015 2015 2015 ...
 $ mon           : num  12 12 12 12 12 12 12 12 12 12 ...
 $ mday          : num  24 24 24 24 24 24 24 24 24 24 ...
 $ hour          : num  0 0 0 0 0 0 0 1 1 1 ...
 $ min           : num  9 19 29 46 49 56 59 16 26 29 ...
 $ tzname        : chr  "America/Sao_Paulo" "America/Sao_Paulo" "America/Sao_Paulo" "America/Sao_Paulo" ...
 $ id            : chr  "IRIODEJA53" "IRIODEJA53" "IRIODEJA53" "IRIODEJA53" ...
 $ latitude      : num  -22.8 -22.8 -22.8 -22.8 -22.8 ...
 $ longitude     : num  -43.2 -43.2 -43.2 -43.2 -43.2 ...
 $ tempm         : num  30.6 30.5 30.5 30.4 30.3 30.3 30.2 30.1 30 30 ...
 $ tempi         : num  87.1 86.9 86.9 86.7 86.5 86.5 86.4 86.2 86 86 ...
 $ dewptm        : num  22.7 22.6 22.6 22.5 22.4 22.4 22.3 22.2 22.3 22.3 ...
 $ dewpti        : num  72.8 72.6 72.6 72.5 72.3 72.3 72.1 72 72.2 72.2 ...
 $ hum           : num  62 62 62 62 62 62 62 62 63 63 ...
 $ wspdm         : num  -1609 -1609 -1609 -1609 -1609 ...
 $ wspdi         : num  -1000 -1000 -1000 -1000 -1000 ...
 $ wgustm        : num  -1607 -1607 -1607 -1607 -1607 ...
 $ wgusti        : num  -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
 $ wdird         : num  -9999 -9999 -9999 -9999 -9999 ...
 $ wdire         : chr  "North" "North" "North" "North" ...
 $ pressurem     : num  1016 1016 1016 1016 1016 ...
 $ pressurei     : num  30 30 30 30 30 ...
 $ windchillm    : num  -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
 $ windchilli    : num  -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
 $ heatindexm    : num  34.5 34.3 34.3 34 33.8 33.8 33.7 33.5 33.4 33.4 ...
 $ heatindexi    : num  94.1 93.7 93.7 93.3 92.8 92.8 92.6 92.2 92.2 92.2 ...
 $ precip_ratem  : num  -2540 -2540 -2540 -2540 -2540 ...
 $ precip_ratei  : num  -100 -100 -100 -100 -100 ...
 $ precip_totalm : num  -2540 -2540 -2540 -2540 -2540 ...
 $ precip_totali : num  -100 -100 -100 -100 -100 ...
 $ solarradiation: chr  "" "" "" "" ...
 $ UV            : chr  "" "" "" "" ...
 $ softwaretype  : chr  "Netatmo" "Netatmo" "Netatmo" "Netatmo" ...
}}
\usage{
Rio_history
}
\description{
This contains the weather history for the Personal Weather Stations in
Rio de Janeiro, Brazil, in the period Dec 24,2015 - Dec, 31 2015.
It is the output of the following call of the 'PWS_history' function,
Rio_history <- PWS_history(Rio_metadata,"20151224","20151231",your.key)
where, your.key' should be replaced with your API-key for Weather Underground
and 'Rio_metadata' is provided in the package (load it via data(Rio_metadata)).
}
\examples{
data(Rio_history)
head(Rio_history)
}
\author{
wundr team
}
\keyword{datasets}

