% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_metadata}
\alias{Rio_metadata}
\title{Rio_metadata dataset}
\format{\preformatted{List of 2
 $ PWSmetadata:'data.frame':	36 obs. of  9 variables:
  ..$ neighborhood: chr [1:36] "Jardim Guanabara" "Santa Teresa" "APA Sao Jose" "Fonseca" ...
  ..$ city        : chr [1:36] "Rio De Janeiro" "Rio De Janeiro" "Maricá" "Niterói" ...
  ..$ state       : chr [1:36] "" "" "RIO DE JANEIRO" "" ...
  ..$ country     : chr [1:36] "BZ" "BZ" "BRASIL" "BZ" ...
  ..$ id          : chr [1:36] "IRIODEJA53" "IRIODEJA63" "IRIODEJA79" "INITERI3" ...
  ..$ lat         : num [1:36] -22.8 -22.9 -22.9 -22.9 -22.9 ...
  ..$ lon         : num [1:36] -43.2 -43.2 -43.2 -43.1 -43.2 ...
  ..$ distance_km : num [1:36] 5.97 6.57 8.4 9.16 9.81 ...
  ..$ distance_mi : num [1:36] 3.71 4.08 5.22 5.69 6.09 ...
 $ call       :List of 3
  ..$ lon      : num -43.2
  ..$ lat      : num -22.9
  ..$ radius_km: num 50
}}
\usage{
Rio_metadata
}
\description{
This is the meta_data for the Personal Weather Stations in Rio de Janeiro, Brazil.
It is the output of the following call of the 'PWS_meta_query' function, where
'your.key' shoudl be replaced with your API-key for Weather Underground:
Rio_metadata <- PWS_meta_query(-22.856878, -43.185368, 50, your.key)
}
\examples{
data(Rio_metadata)
head(Rio_metadata$PWSmetadata)
}
\author{
wundr team
}
\keyword{datasets}

