% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{set_basemap}
\alias{set_basemap}
\title{set_basemap}
\usage{
set_basemap(PWS.class, zoom = 9)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{zoom}{A zoom level for the contextual map. Greater values result in
increased zoom. Default = 9, which captures a 50-mile radius.}
}
\value{
A ggmap raster object.
}
\description{
Creates subtle contextual map background for ggplots based on extent points.
}
\examples{
data(dm_cond)
set_basemap(dm_cond)
}

