% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setConds}
\alias{setConds}
\title{Set the wundr package conds option}
\usage{
setConds(conds)
}
\arguments{
\item{value}{to assign}
}
\value{
the changed value of the package global \code{wundr.options}
}
\description{
Set the wundr package conds option
}
\examples{
\dontrun{
  setConds(data.frame)
}

}

