% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setCount}
\alias{setCount}
\title{Reset the global package variable \code{wundr.options}}
\usage{
setCount(API.count)
}
\value{
the default value value of wundr package global \code{wundr.options}
}
\description{
Reset the global package variable \code{wundr.options}
}
\examples{
\dontrun{
  setCount(numeric)
}

}

