% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setRepository}
\alias{setRepository}
\title{Set the wundr maximum number of iterations option}
\usage{
setRepository(repository)
}
\arguments{
\item{maxIter, }{the value to assign}
}
\value{
the changed value of the package global \code{wundr.options}
}
\description{
Set the wundr maximum number of iterations option
}
\examples{
\dontrun{
  setRepository(NULL)
}

}

