% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{simple_density}
\alias{simple_density}
\title{simple_density}
\usage{
simple_density(PWS.class, title = NULL, add = FALSE, ...)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{title}{A title for you plot. Default = NULL.}

\item{add}{Weather to add this "on top" of previous plot. Default = NULL.}
}
\description{
Creates a simple density raster plot (heatmap) of PWS locations.
}
\examples{
data(dm_cond)
simple_density(dm_cond, "Des Moines, IA PWS Locations")
}

