% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{toSPntsDF}
\alias{toSPntsDF}
\title{toSPntsDF}
\usage{
toSPntsDF(df)
}
\arguments{
\item{df}{A data frame of PWS meta data.}
}
\value{
A spatialPointsDataFrame.
}
\description{
Transform a data frame of PWS meta data to a spatialPointsDataFrame.
}
\examples{
toSPntsDF(Rio_metadata)
}

