% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{webmap_pnts}
\alias{webmap_pnts}
\title{webmap_pnts}
\usage{
webmap_pnts(PWS.class)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}
}
\description{
Interactive web map of PWS stations.
}
\examples{
data(dm_cond)
webmap_pnts(dm_cond)
}

