% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{webmap_raster}
\alias{webmap_raster}
\title{webmap_raster}
\usage{
webmap_raster(PWS.class)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}
}
\description{
Interactive web map of PWS stations density (heatmap).
}
\examples{
data(PWS.Conds.Chicago)
webmap_raster(PWS.Conds.Chicago)
}

