% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.History.R
\docType{data}
\name{Chicago.forecast.10day}
\alias{Chicago.forecast.10day}
\title{Chicago 10-day Forecast}
\format{An object of class \code{list} of length 12.}
\usage{
Chicago.forecast.10day
}
\description{
This contains at 10-day Forecast over all the PWS using wundr's kriging method
}
\examples{
data(Chicago.forecast.10day)
}
\author{
wundr team
}
\keyword{datasets}

