% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartoDB_functions.R
\name{fillblanks}
\alias{fillblanks}
\title{fillblanks}
\usage{
fillblanks(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A data frame.
}
\description{
A helper function used to construct sanitized URLs in CartoDB fucntions.
}
\examples{
fillblanks("foo bar")
}

