% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartoDB_functions.R
\name{get_cdb_table}
\alias{get_cdb_table}
\title{get_cdb_table}
\usage{
get_cdb_table(table_name, cdb_account)
}
\arguments{
\item{table_name}{A PostGres-based table from CartoDB.}

\item{cdb_account}{A CartoDB account name.}
}
\value{
A data frame.
}
\description{
Provides a simple facilty to the CartoDB SQL API
(\url{http://docs.cartodb.com/cartodb-platform/sql-api/}) to import SQL
spatial tabledata from CartoDB into an R data frame. NOTE: you will need to
use your account name and Authorization Key.
}
\examples{
# test CartoDB connection by retrieving existing table ...
\dontrun{stations <- get_cdb_table("public.stations", your.cdb.account)}
}

