% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{gg_points}
\alias{gg_points}
\title{gg_points}
\usage{
gg_points(PWS.class, basemap = basemap, title = NULL, ...)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{basemap}{A contextual basemap. See set_basemap.}
}
\description{
Plots PWS locations on a contextual basemap.
}
\examples{
data(dm_cond)
basemapDM <- set_basemap(dm_cond, zoom = 9)
gg_points(dm_cond)
}

