% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{GP_fit}
\alias{GP_fit}
\title{GP_fit}
\usage{
GP_fit(data.geo, ...)
}
\arguments{
\item{data.geo}{geodata object with coordinates and variable}

\item{...}{Additional parameters to be passed to create_grid, i.e. for changing the grid limitations}
}
\value{
A Gaussian process fit of the variable on all points on a rectangular grid spanning the region
}
\description{
A Gaussian process fit (Kriging) on geodata object.
}
\examples{
data <- Rio_conditions
data.geo <- create_geo_cond(Rio_conditions,"temp_c")
model<-GP_fit(data.geo)
ggplot2::ggplot(data = model, ggplot2::aes(x=lon, y=lat)) +
 ggplot2::geom_tile(ggplot2::aes(fill = value),colour = "white") +
 ggplot2::scale_fill_gradient(low = "yellow", high = "red") +
 ggplot2::geom_point(data=Rio_metadata$PWSmetadata,col='black')
}

