% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Locations.R
\docType{methods}
\name{initialize,PWS.Locations-method}
\alias{initialize,PWS.Locations-method}
\title{An S4 class to represent a bank account.}
\usage{
\S4method{initialize}{PWS.Locations}(.Object, longitude, latitude, radius,
  user.key)
}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{A length-one numeric vector}
}}

