% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Locations.R
\name{j.geocode}
\alias{j.geocode}
\title{j.geocode}
\usage{
j.geocode(address)
}
\arguments{
\item{address}{user character input}
}
\value{
A vector of a (longitude,latitude) pair in numeric format
}
\description{
j.geocode
}
\examples{
j.geocode("Santa Monica, CA")

}

