% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\docType{class}
\name{PWS.Conditions-class}
\alias{PWS.Conditions-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{A length-one numeric vector}

\item{\code{spatialPt}}{A legnth-one numeric vector}

\item{\code{call}}{list of stuff}

\item{\code{data}}{data}
}}

