% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Conditions.R
\docType{data}
\name{PWS.Conds.Chicago}
\alias{PWS.Conds.Chicago}
\title{Chicago Conditions Dataset}
\format{\preformatted{Formal class 'PWS.Conditions' [package "wundr"] with 4 slots
  ..@ spatialPtDF:Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  .. .. ..@ data       :'data.frame':	24 obs. of  9 variables:
  .. .. .. ..$ neighborhood: chr [1:24] "South Loop (Central Station)" "South Loop" "Near South Side" "Near South Side" ...
  .. .. .. ..$ city        : chr [1:24] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. .. .. ..$ state       : chr [1:24] "IL" "IL" "IL" "IL" ...
  .. .. .. ..$ country     : chr [1:24] "US" "US" "US" "US" ...
  .. .. .. ..$ id          : chr [1:24] "KILCHICA310" "KILCHICA388" "KILCHICA315" "KILCHICA166" ...
  .. .. .. ..$ lat         : num [1:24] 41.9 41.9 41.9 41.9 41.9 ...
  .. .. .. ..$ lon         : num [1:24] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..$ distance_km : num [1:24] 1.61 1.99 1.95 2.38 1.03 ...
  .. .. .. ..$ distance_mi : num [1:24] 1.003 1.234 1.213 1.476 0.643 ...
  .. .. ..@ coords.nrs : num(0) 
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ spatialPt  :Formal class 'SpatialPoints' [package "sp"] with 3 slots
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ call       :List of 3
  .. ..$ lon      : num -87.6
  .. ..$ lat      : num 41.9
  .. ..$ radius_km: num 5
  ..@ data       :'data.frame':	24 obs. of  70 variables:
  .. ..$ full                            : chr [1:24] "Chicago, IL" "Chicago, IL" "Chicago, IL" "Chicago, IL" ...
  .. ..$ city                            : chr [1:24] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. ..$ state                           : chr [1:24] "IL" "IL" "IL" "IL" ...
  .. ..$ state_name                      : chr [1:24] "Illinois" "Illinois" "Illinois" "Illinois" ...
  .. ..$ country                         : chr [1:24] "US" "US" "US" "US" ...
  .. ..$ country_iso3166                 : chr [1:24] "US" "US" "US" "US" ...
  .. ..$ zip                             : num [1:24] 60605 60605 60605 60605 60290 ...
  .. ..$ magic                           : num [1:24] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ wmo                             : num [1:24] 99999 99999 99999 99999 99999 ...
  .. ..$ latitude                        : num [1:24] 41.9 41.9 41.9 41.9 41.9 ...
  .. ..$ longitude                       : num [1:24] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. ..$ elevation                       : num [1:24] 182 182 182 182 189 189 190 190 177 178 ...
  .. ..$ display_location.full           : chr [1:24] "Chicago, IL" "Chicago, IL" "Chicago, IL" "Chicago, IL" ...
  .. ..$ display_location.city           : chr [1:24] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. ..$ display_location.state          : chr [1:24] "IL" "IL" "IL" "IL" ...
  .. ..$ display_location.state_name     : chr [1:24] "Illinois" "Illinois" "Illinois" "Illinois" ...
  .. ..$ display_location.country        : chr [1:24] "US" "US" "US" "US" ...
  .. ..$ display_location.country_iso3166: chr [1:24] "US" "US" "US" "US" ...
  .. ..$ display_location.zip            : num [1:24] 60605 60605 60605 60605 60290 ...
  .. ..$ display_location.magic          : num [1:24] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ display_location.wmo            : num [1:24] 99999 99999 99999 99999 99999 ...
  .. ..$ display_location.latitude       : num [1:24] 41.9 41.9 41.9 41.9 41.9 ...
  .. ..$ display_location.longitude      : num [1:24] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. ..$ display_location.elevation      : num [1:24] 182 182 182 182 189 189 190 190 177 178 ...
  .. ..$ station_id                      : chr [1:24] "KILCHICA310" "KILCHICA388" "KILCHICA315" "KILCHICA166" ...
  .. ..$ observation_time                : chr [1:24] "Last Updated on March 13, 5:51 PM CDT" "Last Updated on March 13, 5:51 PM CDT" "Last Updated on March 13, 5:51 PM CDT" "Last Updated on March 13, 5:50 PM CDT" ...
  .. ..$ observation_time_rfc822         : chr [1:24] "Sun, 13 Mar 2016 17:51:58 -0500" "Sun, 13 Mar 2016 17:51:32 -0500" "Sun, 13 Mar 2016 17:51:03 -0500" "Sun, 13 Mar 2016 17:50:36 -0500" ...
  .. ..$ observation_epoch               : num [1:24] 1.46e+09 1.46e+09 1.46e+09 1.46e+09 1.46e+09 ...
  .. ..$ local_time_rfc822               : chr [1:24] "Sun, 13 Mar 2016 17:52:14 -0500" "Sun, 13 Mar 2016 17:52:14 -0500" "Sun, 13 Mar 2016 17:52:14 -0500" "Sun, 13 Mar 2016 17:52:14 -0500" ...
  .. ..$ local_epoch                     : num [1:24] 1.46e+09 1.46e+09 1.46e+09 1.46e+09 1.46e+09 ...
  .. ..$ local_tz_short                  : chr [1:24] "CDT" "CDT" "CDT" "CDT" ...
  .. ..$ local_tz_long                   : chr [1:24] "America/Chicago" "America/Chicago" "America/Chicago" "America/Chicago" ...
  .. ..$ local_tz_offset                 : num [1:24] -500 -500 -500 -500 -500 -500 -500 -500 -500 -500 ...
  .. ..$ weather                         : chr [1:24] "Overcast" "Overcast" "Overcast" "Overcast" ...
  .. ..$ temperature_string              : chr [1:24] "45.9 F (7.7 C)" "42.6 F (5.9 C)" "44.2 F (6.8 C)" "43.2 F (6.2 C)" ...
  .. ..$ temp_f                          : num [1:24] 45.9 42.6 44.2 43.2 47.5 48.7 46.8 46.4 51.3 46.9 ...
  .. ..$ temp_c                          : num [1:24] 7.7 5.9 6.8 6.2 8.6 9.3 8.2 8 10.7 8.3 ...
  .. ..$ relative_humidity               : chr [1:24] "92\%" "100\%" "100\%" "96\%" ...
  .. ..$ wind_string                     : chr [1:24] "Calm" "Calm" "Calm" "Calm" ...
  .. ..$ wind_dir                        : chr [1:24] "WNW" "North" "North" "North" ...
  .. ..$ wind_degrees                    : num [1:24] 290 -9999 -9999 -9999 270 ...
  .. ..$ wind_mph                        : num [1:24] 0 -9999 -9999 -9999 2.7 ...
  .. ..$ wind_gust_mph                   : num [1:24] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ wind_kph                        : num [1:24] 0 0 0 0 4.3 0 0 0 0 0 ...
  .. ..$ wind_gust_kph                   : num [1:24] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ pressure_mb                     : num [1:24] 1005 1005 1005 1005 1005 ...
  .. ..$ pressure_in                     : num [1:24] 29.7 29.7 29.7 29.7 29.7 ...
  .. ..$ pressure_trend                  : chr [1:24] "-" "-" "-" "-" ...
  .. ..$ dewpoint_string                 : chr [1:24] "44 F (7 C)" "43 F (6 C)" "44 F (7 C)" "42 F (6 C)" ...
  .. ..$ dewpoint_f                      : num [1:24] 44 43 44 42 47 47 44 44 46 47 ...
  .. ..$ dewpoint_c                      : num [1:24] 7 6 7 6 8 8 7 7 8 8 ...
  .. ..$ heat_index_string               : chr [1:24] "NA" "NA" "NA" "NA" ...
  .. ..$ heat_index_f                    : chr [1:24] "NA" "NA" "NA" "NA" ...
  .. ..$ heat_index_c                    : chr [1:24] "NA" "NA" "NA" "NA" ...
  .. ..$ windchill_string                : chr [1:24] "46 F (8 C)" "43 F (6 C)" "44 F (7 C)" "43 F (6 C)" ...
  .. ..$ windchill_f                     : chr [1:24] "46" "43" "44" "43" ...
  .. ..$ windchill_c                     : chr [1:24] "8" "6" "7" "6" ...
  .. ..$ feelslike_string                : chr [1:24] "46 F (8 C)" "43 F (6 C)" "44 F (7 C)" "43 F (6 C)" ...
  .. ..$ feelslike_f                     : num [1:24] 46 43 44 43 48 49 47 46 51.3 47 ...
  .. ..$ feelslike_c                     : num [1:24] 8 6 7 6 9 9 8 8 10.7 8 ...
  .. ..$ visibility_mi                   : num [1:24] 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 ...
  .. ..$ visibility_km                   : num [1:24] 2.4 2.4 2.4 2.4 2.4 2.4 2.4 2.4 2.4 2.4 ...
  .. ..$ solarradiation                  : chr [1:24] "6" "--" "--" "--" ...
  .. ..$ UV                              : num [1:24] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ precip_1hr_string               : chr [1:24] "0.07 in ( 2 mm)" "0.00 in ( 0 mm)" "-999.00 in ( 0 mm)" "-999.00 in ( 0 mm)" ...
  .. ..$ precip_1hr_in                   : num [1:24] 0.07 0 -999 -999 -999 0 -999 -999 -999 -999 ...
  .. ..$ precip_1hr_metric               : num [1:24] 2 0 0 0 0 0 0 0 0 0 ...
  .. ..$ precip_today_string             : chr [1:24] "0.65 in (17 mm)" "0.00 in (0 mm)" "-999.00 in (-25375 mm)" "-999.00 in (-25375 mm)" ...
  .. ..$ precip_today_in                 : num [1:24] 0.65 0 -999 -999 0.51 0 -999 -999 -999 -999 ...
  .. ..$ precip_today_metric             : chr [1:24] "17" "0" "--" "--" ...
}}
\usage{
PWS.Conds.Chicago
}
\description{
This contains the weather conditions for the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
called with PWS.Conds.Chicago <- PWS.Conditions(PWS.Loc.Chicago, user.key=jam.key.2)
where the PWS.Loc.Chicago call is described in the PWS.Locations S4 Class
}
\examples{
data(PWS.Conds.Chicago)
}
\author{
wundr team
}
\keyword{datasets}

