<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
% Please edit documentation in R/S4.History.R
\docType{data}
\name{PWS.Hist.Chicago}
\alias{PWS.Hist.Chicago}
\title{Chicago History Dataset}
<<<<<<< HEAD
\format{An object of class \code{PWS.History} of length 1.}
=======
\format{\preformatted{Formal class 'PWS.History' [package "wundr"] with 16 slots
  ..@ average.Temp          :'data.frame':	12 obs. of  2 variables:
  .. ..$ id   : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ tempi: num [1:12] 37.8 35.6 30.6 39.9 40.1 ...
  ..@ average.Humidity      :'data.frame':	12 obs. of  2 variables:
  .. ..$ id : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ hum: num [1:12] 69.8 70.5 93.2 64.1 60.2 ...
  ..@ average.Pressure      :'data.frame':	12 obs. of  2 variables:
  .. ..$ id       : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ pressurei: num [1:12] 30.2 30.3 30.2 30.1 29.8 ...
  ..@ average.Dew.Point     :'data.frame':	12 obs. of  2 variables:
  .. ..$ id    : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ dewpti: num [1:12] 28.4 26.5 28.9 28.6 27.2 ...
  ..@ variance.Temp         :'data.frame':	12 obs. of  2 variables:
  .. ..$ id   : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ tempi: num [1:12] 130.8 130.3 61.8 116.8 107.2 ...
  ..@ variance.Humidity     :'data.frame':	12 obs. of  2 variables:
  .. ..$ id : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ hum: num [1:12] 156.1 150.8 73.2 97.5 90.2 ...
  ..@ variance.Pressure     :'data.frame':	12 obs. of  2 variables:
  .. ..$ id       : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ pressurei: num [1:12] 0.0192 0.0203 0.0166 0.0213 0.0148 ...
  ..@ variance.Dew.Point    :'data.frame':	12 obs. of  2 variables:
  .. ..$ id    : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ dewpti: num [1:12] 119.8 118 99.2 132.2 105.7 ...
  ..@ standard.Dev.Temp     :'data.frame':	12 obs. of  2 variables:
  .. ..$ id   : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ tempi: num [1:12] 11.44 11.41 7.86 10.81 10.35 ...
  ..@ standard.Dev.Humidity :'data.frame':	12 obs. of  2 variables:
  .. ..$ id : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ hum: num [1:12] 12.49 12.28 8.56 9.88 9.49 ...
  ..@ standard.Dev.Pressure :'data.frame':	12 obs. of  2 variables:
  .. ..$ id       : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ pressurei: num [1:12] 0.139 0.142 0.129 0.146 0.121 ...
  ..@ standard.Dev.Dew.Point:'data.frame':	12 obs. of  2 variables:
  .. ..$ id    : chr [1:12] "KILCHICA130" "KILCHICA132" "KILCHICA166" "KILCHICA167" ...
  .. ..$ dewpti: num [1:12] 10.95 10.86 9.96 11.5 10.28 ...
  ..@ history               :'data.frame':	8652 obs. of  34 variables:
  .. ..$ pretty        : chr [1:8652] "12:02 AM CST on March 06, 2015" "12:23 AM CST on March 06, 2015" "12:33 AM CST on March 06, 2015" "12:43 AM CST on March 06, 2015" ...
  .. ..$ year          : num [1:8652] 2015 2015 2015 2015 2015 ...
  .. ..$ mon           : num [1:8652] 3 3 3 3 3 3 3 3 3 3 ...
  .. ..$ mday          : num [1:8652] 6 6 6 6 6 6 6 6 6 6 ...
  .. ..$ hour          : num [1:8652] 0 0 0 0 0 1 1 1 1 1 ...
  .. ..$ min           : num [1:8652] 2 23 33 43 53 3 13 23 33 43 ...
  .. ..$ tzname        : chr [1:8652] "America/Chicago" "America/Chicago" "America/Chicago" "America/Chicago" ...
  .. ..$ id            : chr [1:8652] "KILCHICA166" "KILCHICA166" "KILCHICA166" "KILCHICA166" ...
  .. ..$ latitude      : num [1:8652] 41.9 41.9 41.9 41.9 41.9 ...
  .. ..$ longitude     : num [1:8652] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. ..$ tempm         : num [1:8652] -8.6 -8.8 -8.8 -8.9 -8.9 -9 -9 -9.1 -9.1 -9.2 ...
  .. ..$ tempi         : num [1:8652] 16.5 16.2 16.2 16 16 15.8 15.8 15.6 15.6 15.4 ...
  .. ..$ dewptm        : num [1:8652] -12.1 -12.3 -12.1 -12.4 -12.2 -12.3 -12.3 -12.4 -12.4 -12.3 ...
  .. ..$ dewpti        : num [1:8652] 10.2 9.9 10.2 9.7 10 9.8 9.8 9.6 9.6 9.8 ...
  .. ..$ hum           : num [1:8652] 76 76 77 76 77 77 77 77 77 78 ...
  .. ..$ wspdm         : num [1:8652] -1609 -1609 -1609 -1609 -1609 ...
  .. ..$ wspdi         : num [1:8652] -1000 -1000 -1000 -1000 -1000 ...
  .. ..$ wgustm        : num [1:8652] -1607 -1607 -1607 -1607 -1607 ...
  .. ..$ wgusti        : num [1:8652] -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
  .. ..$ wdird         : num [1:8652] -9999 -9999 -9999 -9999 -9999 ...
  .. ..$ wdire         : chr [1:8652] "North" "North" "North" "North" ...
  .. ..$ pressurem     : num [1:8652] 1036 1035 1035 1035 1035 ...
  .. ..$ pressurei     : num [1:8652] 30.6 30.6 30.6 30.6 30.6 ...
  .. ..$ windchillm    : num [1:8652] -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
  .. ..$ windchilli    : num [1:8652] -999 -999 -999 -999 -999 -999 -999 -999 -999 -999 ...
  .. ..$ heatindexm    : num [1:8652] -9999 -9999 -9999 -9999 -9999 ...
  .. ..$ heatindexi    : num [1:8652] -9999 -9999 -9999 -9999 -9999 ...
  .. ..$ precip_ratem  : num [1:8652] -2540 -2540 -2540 -2540 -2540 ...
  .. ..$ precip_ratei  : num [1:8652] -100 -100 -100 -100 -100 ...
  .. ..$ precip_totalm : num [1:8652] -2540 -2540 -2540 -2540 -2540 ...
  .. ..$ precip_totali : num [1:8652] -100 -100 -100 -100 -100 ...
  .. ..$ solarradiation: chr [1:8652] "" "" "" "" ...
  .. ..$ UV            : chr [1:8652] "" "" "" "" ...
  .. ..$ softwaretype  : chr [1:8652] "Netatmo" "Netatmo" "Netatmo" "Netatmo" ...
  ..@ spatialPtDF           :Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  .. .. ..@ data       :'data.frame':	24 obs. of  9 variables:
  .. .. .. ..$ neighborhood: chr [1:24] "South Loop (Central Station)" "South Loop" "Near South Side" "Near South Side" ...
  .. .. .. ..$ city        : chr [1:24] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. .. .. ..$ state       : chr [1:24] "IL" "IL" "IL" "IL" ...
  .. .. .. ..$ country     : chr [1:24] "US" "US" "US" "US" ...
  .. .. .. ..$ id          : chr [1:24] "KILCHICA310" "KILCHICA388" "KILCHICA315" "KILCHICA166" ...
  .. .. .. ..$ lat         : num [1:24] 41.9 41.9 41.9 41.9 41.9 ...
  .. .. .. ..$ lon         : num [1:24] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..$ distance_km : num [1:24] 1.61 1.99 1.95 2.38 1.03 ...
  .. .. .. ..$ distance_mi : num [1:24] 1.003 1.234 1.213 1.476 0.643 ...
  .. .. ..@ coords.nrs : num(0) 
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ spatialPt             :Formal class 'SpatialPoints' [package "sp"] with 3 slots
  .. .. ..@ coords     : num [1:24, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.8 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ call                  :List of 3
  .. ..$ lon      : num -87.6
  .. ..$ lat      : num 41.9
  .. ..$ radius_km: num 5
}}
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
\usage{
PWS.Hist.Chicago
}
\description{
This contains the weather history for the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
called with PWS.Hist.Chicago <- PWS.History(PWS.Loc.Chicago, "20150306", "20150310", user.key)
where the PWS.Loc.Chicago call is described in the PWS.Locations S4 Class
}
\examples{
data(PWS.Hist.Chicago)
<<<<<<< HEAD
head(PWS.Hist.Chicago)
=======
>>>>>>> 4484b5c1be53e04547ad74ec540c400889c53519
}
\author{
wundr team
}
\keyword{datasets}

