% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.History.R
\name{PWS.History}
\alias{PWS.History}
\title{PWS.Locations constructor function}
\usage{
PWS.History(...)
}
\arguments{
\item{...}{coordinates of initializer function}
}
\value{
S4 class
}
\description{
PWS.Locations constructor function
}

