% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.Locations.R
\docType{data}
\name{PWS.Loc.Chicago}
\alias{PWS.Loc.Chicago}
\title{Chicago Locations Dataset}
\format{An object of class \code{PWS.Locations} of length 1.}
\usage{
PWS.Loc.Chicago
}
\description{
This contains a listing of the Personal Weather Stations in
a 5km region centered in Chicago, Illinois from March 13, 2016
(PWS.Loc.Chicago <- PWS.Locations("Chicago, IL", radius=5, user.key)).
}
\examples{
data(PWS.Loc.Chicago)
head(PWS.Loc.Chicago)
}
\author{
wundr team
}
\keyword{datasets}

