% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Query.Subset.R
\docType{data}
\name{PWS.Loc.Sub.Chicago}
\alias{PWS.Loc.Sub.Chicago}
\title{Chicago History Dataset}
\format{\preformatted{Formal class 'PWS.Query.Subset' [package "wundr"] with 3 slots
  ..@ spatialPtDF:Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  .. .. ..@ data       :'data.frame':	14 obs. of  9 variables:
  .. .. .. ..$ neighborhood: chr [1:14] "Central" "Central" "Central" "Central" ...
  .. .. .. ..$ city        : chr [1:14] "Chicago" "Chicago" "Chicago" "Chicago" ...
  .. .. .. ..$ state       : chr [1:14] "IL" "IL" "IL" "IL" ...
  .. .. .. ..$ country     : chr [1:14] "US" "US" "US" "US" ...
  .. .. .. ..$ id          : chr [1:14] "KILCHICA186" "KILCHICA208" "KILCHICA231" "KILCHICA268" ...
  .. .. .. ..$ lat         : num [1:14] 41.9 41.9 41.9 41.9 41.9 ...
  .. .. .. ..$ lon         : num [1:14] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..$ distance_km : num [1:14] 1.7 1.53 1.93 1.56 1.92 ...
  .. .. .. ..$ distance_mi : num [1:14] 1.054 0.948 1.202 0.969 1.195 ...
  .. .. ..@ coords.nrs : num(0) 
  .. .. ..@ coords     : num [1:14, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.9 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ spatialPt  :Formal class 'SpatialPoints' [package "sp"] with 3 slots
  .. .. ..@ coords     : num [1:14, 1:2] -87.6 -87.6 -87.6 -87.6 -87.6 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..@ bbox       : num [1:2, 1:2] -87.7 41.9 -87.6 41.9
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. .. .. ..$ : chr [1:2] "min" "max"
  .. .. ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. ..@ projargs: chr NA
  ..@ call       :List of 3
  .. ..$ lon      : num -87.6
  .. ..$ lat      : num 41.9
  .. ..$ radius_km: num 2
}}
\usage{
PWS.Loc.Sub.Chicago
}
\description{
This contains subregion locations for the Personal Weather Stations in
a 2km region centered near Chicago, Illinois from March 13, 2016
called with PWS.Loc.Sub.Chicago <- PWS.Query.Subset(PWS.Loc.Chicago, -87.62, 41.88, 2)
where the PWS.Loc.Chicago call is described in the PWS.Locations S4 Class (which is a 5km region).
The subregion lies within the larger region.
}
\examples{
data(PWS.Loc.Sub.Chicago.rda)
}
\author{
wundr team
}
\keyword{datasets}

