% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.Locations.R
\docType{class}
\name{PWS.Locations-class}
\alias{PWS.Locations-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{spatialPtDF}}{A length-one numeric vector}

\item{\code{spatialPt}}{A legnth-one numeric vector}

\item{\code{call}}{list of stuff}
}}

