% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\name{PWS_conditions}
\alias{PWS_conditions}
\title{PWS_conditions}
\usage{
PWS_conditions(PWSmetadata, user_key, stdAPI = TRUE)
}
\arguments{
\item{PWSmetadata}{Meta data object of weather stations (output of 'PWS_meta_query')}

\item{user_key}{Your WUnderground API user key given as a character string}

\item{stdAPI}{A bolean variable indicating whether you have a standard (free) API access (TRUE) which only allows for 10 API calls per minute. The default is the standard API.}
}
\value{
A data frame containing all the weather conditions of the stations in the provided meta data object.
}
\description{
Given an output of 'PWS_meta_query' which contains meta data of personal weather stations in a region
the function 'PWS_conditions' downloads the weather conditions for all stations in the meta data.
}
\examples{
# You can download the current conditions from waether stations in Rio de Janeiro using
# the command below if you replace 'your.key' with your API-key from Weather Underground:
# data(Rio_metadata)
# Rio_conditions <- PWS_conditions(Rio_metadata, your.key)
# if you run the above code with your key the output should be the same as provided here:
data(Rio_conditions)
head(Rio_conditions)
}

