% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{PWS_history}
\alias{PWS_history}
\title{PWS_history}
\usage{
PWS_history(PWSmetadata, begin_YYYYMMDD, end_YYYYMMDD, user_key,
  stdAPI = TRUE)
}
\arguments{
\item{PWSmetadata}{Meta data object of weather stations (output of 'PWS_meta_query')}

\item{begin_YYYYMMDD}{Beginning date in format "YYYYMMDD"}

\item{end_YYYYMMDD}{Ending date in format "YYYYMMDD"}

\item{user_key}{Your WUnderground API user key given as a character string}

\item{stdAPI}{A bolean variable indicating whether you have a standard (free) API access (TRUE) which only allows for 10 API calls per minute. The default is the standard API.}
}
\value{
A data frame containing all the weather conditions of the stations in the provided meta data object.
}
\description{
Given an output of 'PWS_meta_query' which contains meta data of personal weather stations in a region
and a start and an end date for a period of time, the function 'PWS_history' downloads the weather
conditions history for all stations in the meta data. Note that 1 API per stations per day is required.
}
\examples{
# You can download the current conditions from weather stations in Rio de Janeiro using
# the command below if you replace 'your.key' with your API-key from Weather Underground:
\dontrun{
data(Rio_metadata)
Rio_history <- PWS_history(Rio_metadata,"20151224","20151231",your.key)
}
# if you run the above code with your key the output should be the same as provided here:
data(Rio_history)
head(Rio_history)
hist.zoo <- history_zoo(Rio_history,"IRIODEJA53",c("hum","tempm"))
plot(hist.zoo,col='red', main = "Humidity and Temperatur")

}

