% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cartoDB_functions.R
\name{r2cdb}
\alias{r2cdb}
\title{r2cdb}
\usage{
r2cdb(user_key, cdb_account, PWS.Conditions)
}
\arguments{
\item{user_key}{A CartoDB user authorization key.}

\item{cdb_account}{A CartoDB account name.}

\item{PWS.Conditions}{A S4.class PWS.Conditions object of PWS locations and conditions attribute
data.}
}
\value{
NULL. Plus, printed information on status of CartoDB API and URL of
  exported map data on CartoDB service.
}
\description{
Povides facilities to export Weather Underground PWS data to CartoDB's
spatial database hosting and webmap publishing platform through its API
(\url{http://docs.cartodb.com/cartodb-platform/sql-api/}). NOTE: you will need
have an account and Authorization Key with CartoDB. You can create a free
account and get your key at \url{https://cartodb.com/signup}.
}
\examples{
# export PWS locations and conditions data to CartoDB
\dontrun{r2cdb(your.cdb.key, your.cdb.account, PWS.Conditions)}
}

