% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_basemap}
\alias{Rio_basemap}
\title{Rio_basemap object}
<<<<<<< HEAD
\format{\preformatted{ chr [1:1280, 1:1280] "#C2D5BB" "#CEE1C6" "#B7C6AE" "#B2C2AA" ...
 - attr(*, "class")= chr [1:2] "ggmap" "raster"
 - attr(*, "bb")='data.frame':	1 obs. of  4 variables:
  ..$ ll.lat: num -23.3
  ..$ ll.lon: num -43.6
  ..$ ur.lat: num -22.5
  ..$ ur.lon: num -42.7
 - attr(*, "source")= chr "google"
 - attr(*, "maptype")= chr "terrain"
 - attr(*, "zoom")= num 10
}}
=======
\format{An object of class \code{ggmap} (inherits from \code{raster}) with 639 rows and 640 columns.}
>>>>>>> origin/master
\usage{
Rio_basemap
}
\description{
Object of tyoe ggmapraster containing the base map of Rio de Janiero. For visualisation
of example data sets. The base map can be plotted using 'ggmap'
}
\examples{
data(Rio_basemap)
require(ggmap)
ggmap(Rio_basemap)
}
\author{
retrived from google maps
}
\keyword{datasets}

