% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_basemap}
\alias{Rio_basemap}
\title{Rio_basemap object}
\format{\preformatted{ chr [1:639, 1:640] "#FFFFFF" "#A6A6A6" "#070707" "#070707" ...
 - attr(*, "class")= chr [1:2] "ggmap" "raster"
 - attr(*, "bb")='data.frame':	1 obs. of  4 variables:
  ..$ ll.lat: num -23.2
  ..$ ll.lon: num -43.6
  ..$ ur.lat: num -22.4
  ..$ ur.lon: num -42.8
 - attr(*, "source")= chr "stamen"
 - attr(*, "maptype")= chr "toner-lite"
 - attr(*, "zoom")= num 10
}}
\usage{
Rio_basemap
}
\description{
Object of tyoe ggmapraster containing the base map of Rio de Janiero. For visualisation
of example data sets. The base map can be plotted using 'ggmap'
}
\examples{
data(Rio_basemap)
require(ggmap)
ggmap(Rio_basemap)
}
\author{
retrived from google maps
}
\keyword{datasets}

