% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_conditions}
\alias{Rio_conditions}
\title{Rio_conditions dataset}
\format{An object of class \code{data.frame} with 34 rows and 70 columns.}
\usage{
Rio_conditions
}
\description{
This contains the weather conditions for the Personal Weather Stations in
Rio de Janeiro, Brazil, downloaded on March 12, 2016.
It is the output of the following call of the 'PWS_conditions' function,
Rio_metadata <- PWS_meta_query(-22.856878, -43.185368, 50, your.key)
where, your.key' should be replaced with your API-key for Weather Underground
and 'Rio_metadata' is provided in the package (load it via data(Rio_metadata)).
}
\examples{
data(Rio_conditions)
head(Rio_conditions)

}
\author{
wundr team
}
\keyword{datasets}

