% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_history}
\alias{Rio_history}
\title{Rio_history dataset}
\format{An object of class \code{data.frame} with 40202 rows and 34 columns.}
\usage{
Rio_history
}
\description{
This contains the weather history for the Personal Weather Stations in
Rio de Janeiro, Brazil, in the period Dec 24,2015 - Dec, 31 2015.
It is the output of the following call of the 'PWS_history' function,
Rio_history <- PWS_history(Rio_metadata,"20151224","20151231",your.key)
where, your.key' should be replaced with your API-key for Weather Underground
and 'Rio_metadata' is provided in the package (load it via data(Rio_metadata)).
}
\examples{
data(Rio_history)
head(Rio_history)

}
\author{
wundr team
}
\keyword{datasets}

