% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{data}
\name{Rio_metadata}
\alias{Rio_metadata}
\title{Rio_metadata dataset}
\format{An object of class \code{list} of length 2.}
\usage{
Rio_metadata
}
\description{
This is the meta_data for the Personal Weather Stations in Rio de Janeiro, Brazil.
It is the output of the following call of the 'Rio_metadata' function, where
'your.key' shoudl be replaced with your API-key for Weather Underground:
Rio_metadata <- PWS_meta_query(-22.856878, -43.185368, 50, your.key)
}
\examples{
data(Rio_metadata)
head(Rio_metadata$PWSmetadata)

}
\author{
wundr team
}
\keyword{datasets}

