% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{simple_density}
\alias{simple_density}
\title{simple_density}
\usage{
simple_density(PWS.class, title = NULL, add = FALSE, ...)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{title}{A title for you plot. Default = NULL.}

\item{add}{Whether to add this "on top" of previous plot. Default = NULL.}

\item{...}{...}
}
\value{
A RasterLayer.
}
\description{
Creates a simple density raster plot (heatmap) of PWS locations.
}
\examples{
data("PWS.Conds.Chicago")
simple_density(PWS.Conds.Chicago, "Hello World!")
}

