% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{simple_pnts}
\alias{simple_pnts}
\title{simple_pnts}
\usage{
simple_pnts(spatialPtDF, title = NULL, add = FALSE, ...)
}
\arguments{
\item{spatialPtDF}{A SpatialPointsDataFrame.}

\item{title}{A title for you plot. Default = NULL.}
}
\description{
A simple 2D plot of spatial points.
}
\examples{
Rio_spdf <- toWGS84(Rio_metadata)
simple_pnts(Rio_spdf, "Rio PWS Locations")
}

