% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{simple_pnts}
\alias{simple_pnts}
\title{simple_pnts}
\usage{
simple_pnts(PWS.class, title = NULL, add = FALSE, ...)
}
\arguments{
\item{PWS.class}{A PWS.class points S4 object.}

\item{title}{A title for you plot. Default = NULL.}
}
\description{
A simple 2D plot of spatial points.
}
\examples{
data(dm_cond)
simple_pnts(dm_cond, "Des Moines, IA PWS Locations")
}

